refactor code modal
